Add tabular transforms and base tabular dataset. (#50)
* Add tabular transforms and base tabular dataset.

* Move Union outside npt.NDArray Union types.

+++ bolts/data/datasets/utils.py
@@ -48,7 +48,7 @@ __all__ = [
 ImageLoadingBackend = Literal["opencv", "pillow"]
 
 
-RawImage = Union[npt.NDArray[np.int_], Image.Image]
+RawImage = Union[npt.NDArray[np.integer], Image.Image]
 
 
 @overload
